backend/requirements.txt

fastapi
uvicorn
sentence-transformers
pymongo
faiss-cpu
requests
transformers
einops
